seccomp: Put some restrictions on open(2)
We can restrict the open syscall somewhat by only allowing the two log
files to be opened for writing and any other files to be opened for
reading only.

The first case is for log file rotation and we can do that by matching
the addresses of the log file string pointers. The second case is for
things like getpwnam(3) which needs to read /etc/passwd.

Signed-off-by: Andrew Clayton <andrew.clayton@securolytics.io>