Re-work SIGTERM signal handling
In the master process rather than just exiting after sending SIGTERM to
the workers in the SIGTERM signal handler, just set a flag that we got a
SIGTERM signal.

Then back in the main loop we can check for this flag and if it's set we
can then send SIGTERM to the workers and then break out the loop going
through various bits of cleanup and then exiting.

In the worker processes we add SIGTERM to the signalfd() handler, moving
it out to a new function and checking for SIGHUP (log file rotation) or
SIGTERM and acting accordingly doing some cleanup as required and
exiting.

Signed-off-by: Andrew Clayton <andrew.clayton@securolytics.io>