seccomp: Allow getpid(2)
Under Fedora 26 (glibc 2.25) I noticed that when using seccomp getpid
was returning -13, which is -EACCES. We use getpid in the log output.

It used to be that glibc would cache pids and thus you wouldn't actually
see a getpid system call. However this changed with this commit[0], also
see[1].

So we add getpid() to the list of allowed system calls in seccomp.

[0] - http://sourceware.org/git/?p=glibc.git;a=commit;h=c579f48edba88380635ab98cb612030e3ed8691e
[1] - https://sourceware.org/glibc/wiki/Release/2.25#pid_cache_removal

Signed-off-by: Andrew Clayton <andrew.clayton@securolytics.io>