Add SPDX identifiers
The Software Package Data Exchange[0] (SPDX) is a Linux Foundation
initiative to make it easier for people to see exactly what licenses the
software they use come under.

For this we simply add a single line at the top of each source file with
the appropriate header, in this case it is mainly

    /* SPDX-License-Identifier: GPL-2.0 */

with the exception of tls_sni.[ch] which are

    /* SPDX-License-Identifier: BSD-2-Clause */

as these came from the sniproxy[1] project.

As part of this we rename the COPYING (GPL license text) file to
GPL-2.0 (to match the SPDX identifier) and add a BSD-2-Clause file
containing the license text thereof.

[0] - https://spdx.org/
[1] - https://github.com/dlundquist/sniproxy

Signed-off-by: Andrew Clayton <andrew.clayton@securolytics.io>