sprotly: Simplify the RLIMIT_NOFILE log messages
Reduce some duplicated code in the log messages.

This also required enhancing the logit() macro.

We now need to be able to make multiple logit() calls treated as a
single one. i.e don't add the timestamp and pid to each message, only
the first one.

To do this we add a simple mechanism whereby if a logit() string starts
with a space it will be treated as a continuation of the previous
message.

Signed-off-by: Andrew Clayton <andrew.clayton@securolytics.io>