Re-work logging macros
cppcheck was throwing a bunch of messages like

    [proxy.c:280] -> [proxy.c:290]: (style) Local variable 'len' shadows outer variable

This was due to variables in the macros with the same name as variables in
the function they are called from.

Now this could be fixed by renaming things. But really it comes down to
the macros doing too much.

So, the logit macro is fully converted to an inline function while the
logerr macro is mostly changed into an inline function except we wrap it
in a macro call so we can pass __func__ into the function and get the
name of function it's being called from, rather than the name of the
logerr function.

err_exit remains a macro, it's essentially a single fprintf() call and
is doing the __func__ thing so would need to be wrapped like logerr.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>