proxy: Protect against epoll_wait(2) hangs from starvation
This a continuation of the work to address the issues raised in the
below linked GH issue.

This attempts to address an issue where for whatever reason we could get
stuck in epoll_wait(2) waiting to be notified of events which we never
get despite fd's being ready to be read from or written to.

To do this we employ a dynamic timeout to epoll_wait(2), we start out at
10ms, if we timeout and there is still nothing ready to be read/written,
e.g everything is EAGAIN, then we increase the timeout, under this
condition we will keep increasing the timeout in various steps up to a
maximum of 5s. Whenever we do do a read/write we reset the timeout back
to 10ms.

Link: https://github.com/ac000/sprotly/issues/2
Reported-by: grzech0 <https://github.com/grzech0>
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>